add: progress on utility bar logic
